emacs: add me/setup-language-package macro
This new macro acts as a small wrapper around `use-package`,
and will make it so that use-package is only called when we
try to edit a file containing the provided file extension.

This macro isn't meant to supersede `use-package`, but provide
a way to make it easier to have syntax and editing support
for languages that I don't use often enough to have a dedicated
configuration for.